fix: currency domain expansion
